+++ CHANGELOG.md
@@ -516,3 +516,4 @@ This update emphasizes improving the reliability of Predator, especially when op
         - Previously, the time offset manager thread would keep Predator alive after using the "Quit" option.
     - All threads now watch a global variable, and will exit when it is changed.
         - This means the user can simply press Ctrl+C once, and all threads will clean up and exit.
+- Predator now saves the complete loaded alert database to the `hotlist.json` file in the interface directory.